multi-user: $user may not be provided, authkeys not to be written
Fedora wants to use this in a multi-user fashion.  This means each
"gitolite" user is also a Unix user in his own right.  They will use the
forcecommand mechanism to run gl-auth-command but that's about it.

This changes a few things, and this patch is the first one in that
direction.

  - $user is no longer a mandatory argument to gl-auth; if not given it
    defaults to $USER now
  - writing authkeys no longer makes sense when there is no "central"
    user whose authkeys file needs to be updated.
  - even creating new repos will be done offline, including setting up
    hooks

Note: they have their own mechanism for updating people's authkeys when
they add a developer.