big-config: bypass update hook for people coming through shell
Fedora has admins who (a) have only one key, and (b) need to use git as
well as the shell.  However, this breaks their ability to push.

Disable the update hook if they came through normal ssh+git (instead of
through gl-auth-command).  This allows such users to use the repo (with
full path, of course) as if gitolite didn't exist.